services/simulation/requirements.txt
# FastAPI + server
fastapi==0.110.0
uvicorn[standard]==0.29.0

# OpenTelemetry core and SDK with logging support
opentelemetry-api==1.23.0
opentelemetry-sdk[logs]==1.23.0

# OTLP exporters (HTTP/protobuf)
opentelemetry-exporter-otlp==1.23.0

# Instrumentation packages
opentelemetry-instrumentation==0.44b0
opentelemetry-instrumentation-fastapi==0.44b0
opentelemetry-instrumentation-httpx==0.44b0
opentelemetry-instrumentation-pymongo==0.44b0

# HTTP client
httpx==0.27.0

# MongoDB async client
motor==3.3.1
pymongo[srv]==4.6.1
